safari-600.7-branch/LayoutTests/js/dom/post-message-numeric-property-expected.txt

Tests that posting a message that consists of an object with a numeric property doesn't try to reify the property as a named property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.data[0] is 'hi'
PASS successfullyParsed is true

TEST COMPLETE
